Move java style information into the linting rules page
A lot of this information was already there, but we now have
two suggested checkstyle configs.

This simplies the java page and means we can explain some of the
rationale in a language agnostic way.
I've also linked to a great blog post from GOV.UK explaining the
benefits of setting up automated linting.